archive/requirements_old.txt
aiofiles==24.1.0
aiohttp==3.12.14
capstone==5.0.6
chromadb==1.0.18
docutils
drgn==0.0.32
fastapi==0.116.1
GitPython==3.1.45
haikunator==2.1.0
ioctl_opt==1.3
Jinja2==3.1.6
lark==1.2.2
libevdev==0.12
lxml==6.0.0
mcp==1.13.0
numpy==2.3.2
openai==1.54.0
paramiko==3.5.1
ply==3.11
pydantic==2.11.7
pyelftools==0.32
pyroute2==0.9.4
pytest==8.3.5
python-dotenv==1.1.1
python_jose==3.4.0
pyudev==0.24.3
PyYAML==6.0.2
Requests==2.32.4
requests_unixsocket==0.4.1
scapy==2.6.1
setproctitle==1.3.6
setuptools==78.1.1
sphinx_rtd_dark_mode==1.3.0
sphinx_rtd_theme==3.0.2
thread==2.0.5
Unidecode==1.4.0
uvicorn==0.35.0
